Migration Tab UX
Implemented additional checks/warnings presented with the input/output directories so that issues are more clear to the user if they are to arise. When output is cleared, it will revert to the default directory. Input and output are both checked to make sure they are valid locations and that a limapp is present in the input folder.